+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Added basic support for `TypedArray`s and `Set` values in `struct`, `signature` and `table` views
+- Fixed displaying supported syntaxes in view's showcase for `source` view
 
 ## 1.0.0-beta.80 (21-01-2024)